civic-idea-browser/backend/requirements/base.txt
# Django and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
django-oauth-toolkit==2.2.0

# File Storage & Media
Pillow==10.0.1
django-storages==1.14.2
boto3==1.29.3

# Task Queue & Caching
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# API Documentation
drf-spectacular==0.26.5

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Code Quality
black==23.9.1
flake8==6.1.0
isort==5.12.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3 